app/requirements.txt
mplfinance==0.12.7a17
pandas==1.2.4
pandas-datareader==0.9.0
python-dateutil==2.8.1
python-dotenv==0.17.1
six==1.15.0
urllib3== 1.26.4
pytz==2021.1
pyparsing==2.4.7
Pillow==8.2.0
kiwisolver==1.3.1
idna==2.10
Cycler==0.10.0
chardet==4.0.0
certifi==2020.12.5
matplotlib==3.4.1
lxml==4.6.3
slack-sdk==3.5.1
numpy==1.19.0
